move recv skipper to scheduler_components
